GameStore/requirements.txt
Django==1.9
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.3.0
psycopg2==2.6
static3==0.5.1
python-social-auth==0.2.21
social-auth-app-django
django-tastypie==0.13.3